+++ src/cmd/mod.rs
@@ -2,8 +2,8 @@
 
 pub mod dep;
 pub mod ena;
-pub mod kunitigs;
 pub mod merge;
 pub mod quorum;
 pub mod template;
 pub mod trim;
+pub mod unitigs;