+++ CHANGELOG.md
@@ -11,7 +11,7 @@
   * 4_down_sampling.sh
   * 6_down_sampling.sh
 
-* Add `anchr kunitigs`
+* Add `anchr unitigs`
 
 ## 0.1.5 - 2020-10-24